# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ABM Swagger RestFul API UI",
    "description": "Node.js RestFul API UI",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/match/application": {
      "get": {
        "tags": ["match"],
        "description": "유저가 신청한 매칭 결과들을 parameter의 status를 보고 판단하여 불러온다",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string (JWT token)"
            }
          },
          {
            "name": "status",
            "in": "path",
            "schema": {
              "type": "string",
              "example": "Accepted/Pending/Rejected"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "feedlist": [
                    [
                      {
                        "_id": "63c50102d3e134a6e7cbca73",
                        "title": "asdfasdsdjldsfjljkjaf",
                        "content": "asdfasdsdjldsfjljkjafasdfasdsdjldsfjljkjafasdfasdsdjldsfjljkjafasdfasdsdjldsfjljkjaf",
                        "author": "hehehehehehehehehe",
                        "categories": ["9", "3"],
                        "regions": [
                          {
                            "region_code": "4687000000",
                            "region": "전라남도",
                            "district": "영광군"
                          }
                        ],
                        "state": "Recruiting",
                        "date": "2022-01-07T00:00:00.000Z",
                        "TO": 3,
                        "regularity": true,
                        "thumbnail": "https://ourb.s3.ap-northeast-2.amazonaws.com/77717355455234301.png",
                        "fid": 46,
                        "__v": 0
                      }
                    ],
                    [
                      {
                        "_id": "63c50268a0ff5e4980e88120",
                        "title": "안녕하세요 배드민턴 소모임 모집중입니다",
                        "content": "sad54654654648489fasdf4d8asf984da48f4das465fdas456fa4ds6fas45f5fdsa46fsa456",
                        "author": "asdnfljnnfads",
                        "categories": ["9", "7", "1"],
                        "regions": [
                          {
                            "region": "경기도",
                            "district": "안양시 만안구",
                            "region_code": "4117100000"
                          },
                          {
                            "region": "경기도",
                            "district": "부천시",
                            "region_code": "4119000000"
                          }
                        ],
                        "state": "Recruiting",
                        "date": "2023-01-17T00:00:00.000Z",
                        "TO": 3,
                        "regularity": true,
                        "thumbnail": "https://ourb.s3.ap-northeast-2.amazonaws.com/591167637227303.jpg",
                        "fid": 42,
                        "__v": 0
                      }
                    ]
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/match/application/{fid}": {
      "post": {
        "tags": ["match"],
        "description": "게시글에 match 신청을 넣는 API",
        "parameters": [
          {
            "name": "fid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string (JWT token)"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "DB error"
          },
          "401": {
            "description": "err: re-apply only possible once."
          },
          "402": {
            "description": "state error : Only one application can exist."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/match/recruit/{fid}/{app_user}": {
      "patch": {
        "tags": ["match"],
        "description": "",
        "parameters": [
          {
            "name": "fid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app_user",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "DB Error"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accept": {
                    "example": "Accepted"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}